Update App.js
Using the localStorage of the browser to store the most recent searchTerm so the app opens it in the browser whenever it resarts.

handleSearch uses the stored value (if there is one) as the initial state if the searchTerm, otherwise it defaults to the initial state of 'React'.

Added useEffect hook to ensure that if setSearchTerm is used anywhere else in the app, the localStorage is also updated and stops the app from breaking.